Refac files and functions structure - Resolve bugs: edit on profile-feed(personal-feed), edit post(home), logout(home and profile)
